Simplify OAuth token refresh, unify expiry constants, add JWKS caching
