+++ src/dbt_mcp/oauth/refresh.py
@@ -1,0 +1,33 @@
+import logging
+
+from authlib.integrations.requests_client import OAuth2Session
+
+from dbt_mcp.oauth.client_id import OAUTH_CLIENT_ID
+from dbt_mcp.oauth.dbt_platform import (
+    DbtPlatformContext,
+    dbt_platform_context_from_token_response,
+)
+
+logger = logging.getLogger(__name__)
+
+
+def refresh_oauth_token(
+    *,
+    refresh_token: str,
+    token_url: str,
+    dbt_platform_url: str,
+) -> DbtPlatformContext:
+    """Perform an OAuth token refresh and return the new context.
+
+    This is the single implementation used by both the startup refresh
+    path and the OAuthTokenProvider's inline/background refresh.
+    """
+    oauth_client = OAuth2Session(
+        client_id=OAUTH_CLIENT_ID,
+        token_endpoint=token_url,
+    )
+    token_response = oauth_client.refresh_token(
+        url=token_url,
+        refresh_token=refresh_token,
+    )
+    return dbt_platform_context_from_token_response(token_response, dbt_platform_url)
