+++ src/dbt_mcp/oauth/token.py
@@ -19,11 +19,25 @@ class DecodedAccessToken(BaseModel):
     decoded_claims: dict[str, Any]
 
 
+_jwks_clients: dict[str, PyJWKClient] = {}
+
+
+def _get_jwks_client(jwks_url: str) -> PyJWKClient:
+    if jwks_url not in _jwks_clients:
+        _jwks_clients[jwks_url] = PyJWKClient(jwks_url)
+    return _jwks_clients[jwks_url]
+
+
+def _clear_jwks_cache() -> None:
+    """Clear the cached JWKS clients. Intended for test teardown."""
+    _jwks_clients.clear()
+
+
 def fetch_jwks_and_verify_token(
     access_token: str, dbt_platform_url: str
 ) -> dict[str, Any]:
     jwks_url = f"{dbt_platform_url}/.well-known/jwks.json"
-    jwks_client = PyJWKClient(jwks_url)
+    jwks_client = _get_jwks_client(jwks_url)
     signing_key = jwks_client.get_signing_key_from_jwt(access_token)
     claims = jwt.decode(
         access_token,
