+++ tests/unit/oauth/test_credentials_provider.py
@@ -37,13 +37,13 @@ class TestCredentialsProviderAuthenticationMethod:
 
         with (
             patch(
-                "dbt_mcp.config.settings.get_dbt_platform_context",
+                "dbt_mcp.config.credentials.get_dbt_platform_context",
                 return_value=mock_dbt_context,
             ),
             patch(
-                "dbt_mcp.config.settings.get_dbt_host", return_value="cloud.getdbt.com"
+                "dbt_mcp.config.credentials.get_dbt_host", return_value="cloud.getdbt.com"
             ),
-            patch("dbt_mcp.config.settings.OAuthTokenProvider") as mock_token_provider,
+            patch("dbt_mcp.config.credentials.OAuthTokenProvider") as mock_token_provider,
             patch("dbt_mcp.config.settings.validate_dbt_cli_settings", return_value=[]),
         ):
             mock_provider_instance = MagicMock()
@@ -141,13 +141,13 @@ class TestCredentialsProviderOAuthDoesNotSetDbtToken:
 
         with (
             patch(
-                "dbt_mcp.config.settings.get_dbt_platform_context",
+                "dbt_mcp.config.credentials.get_dbt_platform_context",
                 return_value=mock_dbt_context,
             ),
             patch(
-                "dbt_mcp.config.settings.get_dbt_host", return_value="cloud.getdbt.com"
+                "dbt_mcp.config.credentials.get_dbt_host", return_value="cloud.getdbt.com"
             ),
-            patch("dbt_mcp.config.settings.OAuthTokenProvider") as mock_tp_cls,
+            patch("dbt_mcp.config.credentials.OAuthTokenProvider") as mock_tp_cls,
             patch("dbt_mcp.config.settings.validate_dbt_cli_settings", return_value=[]),
         ):
             mock_tp_cls.create = AsyncMock(return_value=MagicMock())
@@ -181,13 +181,13 @@ class TestCredentialsProviderOAuthDoesNotSetDbtToken:
 
         with (
             patch(
-                "dbt_mcp.config.settings.get_dbt_platform_context",
+                "dbt_mcp.config.credentials.get_dbt_platform_context",
                 return_value=mock_dbt_context,
             ),
             patch(
-                "dbt_mcp.config.settings.get_dbt_host", return_value="cloud.getdbt.com"
+                "dbt_mcp.config.credentials.get_dbt_host", return_value="cloud.getdbt.com"
             ),
-            patch("dbt_mcp.config.settings.OAuthTokenProvider") as mock_tp_cls,
+            patch("dbt_mcp.config.credentials.OAuthTokenProvider") as mock_tp_cls,
             patch("dbt_mcp.config.settings.validate_dbt_cli_settings", return_value=[]),
         ):
             mock_provider_instance = MagicMock()
@@ -225,13 +225,13 @@ class TestCredentialsProviderOAuthDoesNotSetDbtToken:
 
         with (
             patch(
-                "dbt_mcp.config.settings.get_dbt_platform_context",
+                "dbt_mcp.config.credentials.get_dbt_platform_context",
                 return_value=mock_dbt_context,
             ),
             patch(
-                "dbt_mcp.config.settings.get_dbt_host", return_value="cloud.getdbt.com"
+                "dbt_mcp.config.credentials.get_dbt_host", return_value="cloud.getdbt.com"
             ),
-            patch("dbt_mcp.config.settings.OAuthTokenProvider") as mock_tp_cls,
+            patch("dbt_mcp.config.credentials.OAuthTokenProvider") as mock_tp_cls,
             patch("dbt_mcp.config.settings.validate_settings") as mock_validate,
             patch("dbt_mcp.config.settings.validate_dbt_cli_settings", return_value=[]),
         ):
